refactor(entities-plugins): set path alias for src and freeform
