+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/NodeFormJq.vue
@@ -23,12 +23,12 @@
 </template>
 
 <script setup lang="ts">
-import Form from '../../../shared/Form.vue'
-import Field from '../../../shared/Field.vue'
-import InputsField from '../fields/InputsField.vue'
-import { JqNodeSchema } from './mock'
-import useI18n from '../../../../../composables/useI18n'
-import StringField from '../../../shared/StringField.vue'
+import Form from '@freeform/Form.vue'
+import Field from '@freeform/Field.vue'
+import InputsField from './InputsField.vue'
+import { JqNodeSchema } from './schemas'
+import useI18n from '@/composables/useI18n'
+import StringField from '@freeform/StringField.vue'
 
 const { i18n } = useI18n()
 </script>
