+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/OutputValueField.vue
@@ -62,12 +62,12 @@
 // todo(zehao): format user inputs to json before update store, depend on the lazy input sync mode
 import { useTemplateRef, nextTick } from 'vue'
 import { AddIcon, CloseIcon } from '@kong/icons'
-import useI18n from '../../../../../composables/useI18n'
-import { useKeyValueField } from '../../../shared/headless/useKeyValueField'
+import useI18n from '@/composables/useI18n'
+import { useKeyValueField } from '@freeform/headless/useKeyValueField'
 import type {
   KeyValueFieldProps,
   KeyValueFieldEmits,
-} from '../../../shared/headless/useKeyValueField'
+} from '@freeform/headless/useKeyValueField'
 
 const props = defineProps<KeyValueFieldProps>()
 const emit = defineEmits<KeyValueFieldEmits>()
